anodyne/requirements.txt
amqp==2.4.0
argh==0.26.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
Django==2.1.5
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-loginas==0.3.4
django-markdown-deux==1.0.5
django-smtp-ssl==1.0
django-tables2==2.0.4
djangorestframework==3.9.1
Flask==1.0.2
geographiclib==1.49
geopy==1.18.1
googlemaps==3.0.2
haversine==2.0.0
idna==2.8
inotify==0.2.10
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kombu==4.2.2.post1
markdown2==2.3.7
MarkupSafe==1.1.0
nbformat==4.4.0
nose==1.3.7
pathtools==0.1.2
Pillow==5.4.1
plotly==4.1.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycrypto==2.6.1
pytz==2018.9
PyYAML==3.13
reportlab==3.5.13
requests==2.21.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
virtualenv==16.3.0
watchdog==0.9.0
Werkzeug==0.14.1
XlsxWriter==1.1.2
django-queryset-csv
pandas
djangorestframework-simplejwt
django-cors-headers
